requirements.txt
pyyaml>=6.0
click>=8.0.0
colorama>=0.4.6
flask>=2.3.0
flask-cors>=4.0.0

